remote: configurable exponential-backoff retry count
Exponential-backoff retry count can now be configured on the commandline
or in options by flag `--retry-count`. Using a count < 0 returns to the
default fallback of `MaxFailedRetryCount`, currently 20.

e.g

```shell
$ drive push --retry-count 4 a/bc/def terms
```

```shell
$ drive pull --retry-count 20 f1/f2 jk/lm
```

- To use the default retry count:

```shell
$ drive pull --retry-count -1 f1/f2 jk/lm
```

Fixes #529.

+++ src/misc.go
@@ -58,7 +58,7 @@ const (
 )
 
 const (
-	MaxFailedRetryCount = uint32(20) // Arbitrary value
+	MaxFailedRetryCount = 20 // Arbitrary value
 )
 
 var DefaultMaxProcs = runtime.NumCPU()
